packages/agents/idp-agent/pyproject.toml
[project]
name = "idp_v2.idp_agent"
version = "1.0.0"
description = "IDP Agent"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "aws-opentelemetry-distro>=0.14.0",
  "bedrock-agentcore>=0.1.7",
  "fastapi>=0.123.0",
  "boto3>=1.42.0",
  "mcp>=1.22.0",
  "pydantic-settings>=2.0.0",
  "strands-agents>=1.22.0",
  "strands-agents-tools>=0.2.16",
  "uvicorn>=0.38.0"
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [ "." ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[dependency-groups]
dev = [ "fastapi[standard]==0.123.0", "boto3-stubs[dynamodb]" ]
